requirements.txt
arabic-reshaper==2.1.3
hazm==0.7.0
matplotlib==3.5.3
python-bidi==0.4.2
wordcloud==1.8.2.2
loguru==0.6.0
